crctab: use stdint types

+++ crctab.h
@@ -11,8 +11,8 @@
  * wow ! a whole macro ! lets copyright it.....
  */
 
-extern unsigned short int crc16tab[0x100];
-extern unsigned long crc32tab[0x100];
+extern uint16_t crc16tab[0x100];
+extern uint32_t crc32tab[0x100];
 
 #define UPDCRC16(cp, crc) (crc16tab[(((crc) >> 8) & 255)] ^ ((crc) << 8) ^ (cp))
